\name{fasterElasticNet-package}
\alias{fasterElasticNet-package}
\alias{fasterElasticNet}
\docType{package}
\title{
\packageTitle{fasterElasticNet}
}
\description{
\packageDescription{fasterElasticNet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fasterElasticNet}
\packageIndices{fasterElasticNet}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fasterElasticNet}

Maintainer: \packageMaintainer{fasterElasticNet}
}
\references{
~~ Literature or other references for background information ~~
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
}
\keyword{ package }
