\name{fasterElasticNet-package}
\alias{fasterElasticNet-package}
\alias{fasterElasticNet}
\docType{package}
\title{
\packageTitle{fasterElasticNet}
}
\description{
\packageDescription{fasterElasticNet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fasterElasticNet}
\packageIndices{fasterElasticNet}
FasterElasticNet uses some math algorithm such as cholesky decomposition and forward solve etc. to reduce the amount of computation. We also use Rcpp with Armadillo to improve our algorithm by speeding up almost 5 times compared by the R version.
}
\author{
\packageAuthor{"JingyiMa"}

\packageMaintainer{CUFESAM}
}

\keyword{ package }
